cleaned project from unused files and imports

+++ include/Scene/DistantLight.h
@@ -2,7 +2,6 @@
 #define DISTANTLIGHT_H
 
 #include "Light.h"
-#include "limits.h"
 #include <QImage>
 //despite having position because of inheritance DistantLight<-Light<-Object3D
 //its position should never be accessed due to definiton of distant light only its direction should ever be used
@@ -26,8 +25,8 @@ public:
 
     void printShadowMatrix(){
         QImage img(shadowMap.getCols(), shadowMap.getRows(), QImage::Format_Grayscale8);
-        for (int y = 0; y < shadowMap.getRows(); ++y) {
-            for (int x = 0; x < shadowMap.getCols(); ++x) {
+        for (int y = 0; y < (int)shadowMap.getRows(); ++y) {
+            for (int x = 0; x < (int)shadowMap.getCols(); ++x) {
                 float d = shadowMap[y][x];
                 // 0 = close, 1 = far, inf = empty
                 int v = (d == std::numeric_limits<float>::infinity()) ? 0 : int(255 * (1.0 - d));
